Made side panels observe state of JASidePanelController
This way, the side panels will call viewWillAppear and its ilk at the
right moment.  Changing the JASidePanelController to manually send
appearance transition messages to its child view controllers turned out
not to be such a good solution (which I'm sure is why young JA made the
KVO-compliant state field to begin with).