Building some validation into EVExchange.
Made isValid a property on EVObject.  Built self-valildation into
EVExchange and wrote some tests to test it.  Was going to make the
Exchangeable person object also do isValid, but that's just a protocol,
so I don't think it will work in the same way.  Could just rely on
EVUser validation, but you need a lot less info for an Exchangeable
User, so might just not validate it.