Additional work on deposit
Added blue button to the deposit interface and got it updating along
with the deposit amount, etc.

Additionally, I got sidetracked into a k-hole over spreading text
out a little bit and ended up adding EVSpreadLabel to the project,
and I would have pulled it out into a separate branch but it was
already all commingled and fucked, so, you know, sorry. :(