Fixes bugs where cached user is not being properly updated.
A couple of related issues were causing the "me" user not to be
properly updated, particularly as regards the balance.  This commit
adds reload calls upon app launch and willEnterForeground, removes
a shortcut initializer that was preventing the "me" user from ever
being updated, and adds some general code sprinklings to prevent
caching in the future.