Lots of changes, though mostly having to do with the status bar.
Got it working well!  It's very easy to drop in, and it works very well,
though if dismissing a modal controller, you need to do the dismissing
in a completion block passed to the status bar manager.  This is because
of constant sizing the side panel controller does.  I put the status bar
in everything I could think of, other than banks/cards.  Profile needs
to animate more smoothly when editing still.