Just making a couple slight tweaks before making a build.
Fixing a timing issue with the logo coloring, where it could happen
before the status bar completion if the manager had to wait a little.
Also fixed the drop shadow getting laid on top of the status bar when
animating out the success bar.  Also tweaked the layout of the forgot
password link.