Making a little more obvious if you haven't filled out sign up form.
Previously, the save button wouldn't enable till the form was fully
complete.  However, this was confusing if, for example, you entered a
password that was only 7 characters, as it was not immediately obvious
what wouldn't let you sign up.  Now, the button will enable if all the
forms have text, but you will get an alert if you either have a password
less than 7 characters or haven't accepted the TOS/privacy policy.  Also
changed the sizing on the navBar badge when there's more than 1 digit,
it was bothering me.