Tweaking a few problems.
Added the swipe recognizer to self.view instead of
self.navigationController.view.  Hoping this might fix the sporadic
crash we've gotten from the swipeRecognizer.  Also removing the
automaticPropertyValidation - we haven't been using it, and RAC is
pretty processor intensive, so hoping it might speed up the app on the
iPhone 4, which was getting hung up on the main thread upon app launch.
Also put in the wallet pull to refresh, though I need to center it.