Being smarter about loading images.
Previously, if you tried to load the same image multiple times, it would
do the request and image processing each time.  This could be a big
problem when coming into the newsfeed without a cache, and was making it
relatively easy for me to crash on my iPhone 4S by scrolling up and
down.  Now, it will save the success blocks and only perform one
imageRequest for each URL and size combo, running all the blocks at the
end.