Changing Facebook to use its old method for connecting.
The new way uses the native dialog, which is nice, but buggy.  I noticed
almost all my apps just use the old way, sending the user to the
Facebook app, and it seems to work a lot better, though it is slightly
more cumbersome.  It is definitely worth it, though, if it gets rid of
the bugs that are tripping people up.  Another plus is that if the user
declines to give us permission, but then changes their mind, they can
simply hit continue again, whereas before they had to go into settings
to facebook and turn us on.