Making a few changes regarding signup.
Making sure the facebook manager clears its info on sign out, otherwise
there could be a bug when trying to connect again.  Also removing the
loading indicator when connecting, as it doesn't make sense anymore with
the new flow.  Also changing it so the image saved when editing profile
or signing up is 200 pixels small for one of the sides (and proportional
for the other).