Making EVUser parsing use a model mapping method, ala #247.
Same idea as Mantle, though it seems a lot simpler/more readable to me.
I did this fairly quickly, though, so it's definitely possible I'm
missing something.  This shouldn't actually be merged in without a ton
more testing, I was just trying to see if this would tag to the commit
so you could see the code.