Getting the pull to refresh and infinite scrolling right.
This was a nightmare.  Was having all kinds of issues with the content
insets...... not really sure what changes between ios6 and now that
caused it all, but finally managed rto get it straightened out.  Also
fixed the weird spacing at the bottom of the tableView on the home
screen that had been bugging me.