Lots of changes, mostly relating to story cells and navBars.
Getting the story cells to look like our current app.  Kinda tricky
given that ios7 grouped tableViews extend all the way to the side.  Was
also changing all of our 1 point dividers to be 1 pixel, so 0.5 points
on retina.  The Apple stock apps all do this, and it's a sleeker look.
Also was fixing some navBar buttons, which need to be inset differently
in ios7.  Got the main ones, though I think I still need to mess with
some more.