Fixed bug in group record view; minor tweaks.
The main thing in this commit is closing the whole request when you try to
cancel the last person in a group request.

In addition, added a line in the default stylesheet in EVStory in the hopes of
solving our weird boldfacing issue. Also cleaned up an unnecessary log in EVAppDelegate.