Added Facebook friend count logic
First off, added a quiet FB session reopening call to the FacebookManager,
though I don't think it actually works in our current incarnation.

Upon logging in to FB, we now have a call to the friends list and store the
friend count in NSUserDefaults and [EVCIA me], for use in our main menu Invite
cell.

Also refactored a key name to include "Key".