Making sure the app looks right on ios6.
This basically involved going through and adding a bunch of if
statements in places where I had changed layout code.  I still need to
do this for HistoryItems and Invite controllers, the latter of which
still needs to be fixed for ios7.  Both of those had changes in dev, so
I'll need to merge first.